feat: terms i18n
