+++ apps/webapp/app/[lang]/terms/page.tsx
@@ -5,11 +5,14 @@ import {
   PageContentData,
   ContentTextType
 } from '@/components/shared/content'
+import { getDictionary } from '@/app/dictionaries'
+import { CommonPageProps } from '@/app/types/routing.type'
 
-export default function PrivacyPolicy() {
+export default async function PrivacyPolicy({ params }: CommonPageProps) {
+  const dict = await getDictionary(params.lang)
   return (
-    <div className="content-container !py-10 md:px-3 px-7 md:py-24">
-      <PageContent data={content} />
+    <div className="content-container !py-10 px-7 md:px-3 md:py-24">
+      <PageContent data={dict.terms.content} />
     </div>
   )
 }
@@ -19,47 +22,3 @@ export const metadata: Metadata = {
   description:
     'Read our Privacy Policy to understand how we protect and manage your data in our crypto launchpad application.'
 }
-
-const policiesAndTerms = [
-  {
-    title: 'Privacy Policy',
-    content:
-      'We respect your privacy. No personal data is stored or shared with third parties. We only collect emails for our newsletter subscribers and ensure their protection against unauthorized access.'
-  },
-  {
-    title: 'Terms of Service',
-    content:
-      'By using our services, you agree to our terms. Ensure you understand the risks associated with digital currencies and platforms. We do not store your personal data except for email subscriptions.'
-  },
-  {
-    title: 'No Cookie Policy',
-    content:
-      'Our website does not use cookies. Your navigation and interaction with our services are private and not tracked by cookies.'
-  },
-  {
-    title: 'Newsletter Subscription',
-    content:
-      'Subscribers to our newsletter agree to provide their email addresses for regular updates. You can unsubscribe at any time through the link provided in each email.'
-  },
-  {
-    title: 'Data Protection',
-    content:
-      'We implement rigorous security measures to protect your information. Your email is securely stored and is only used for sending newsletters.'
-  }
-] as const
-
-// Define the content using mapped policies and terms
-const content: PageContentData = [
-  {
-    type: 'h1',
-    text: 'Privacy Policy and Terms of Service for Bitlauncher Participants'
-  },
-  {
-    type: 'p',
-    text: 'Understand our commitment to your privacy and your responsibilities when using our crypto launchpad services:'
-  },
-  ...policiesAndTerms.flatMap((item, index) => [
-    { type: 'h2' as ContentTextType, text: `${index + 1}. ${item.title}` },
-    { type: 'p' as ContentTextType, text: item.content }
-  ])
-]
